feat: add support for marking all channels as read
This implements functionality for marking all channels as read. It provides a usecase at the domain level for marking as read. The usecase will optimistically update the UI for marking all as read, and subsequently invoke the client api to trigger backend events. The backend events trigger offline db updates and refreshing the UI if necessary.