feat: CAS-364 implement channel list items for redesign
This is a functioning re-implementation of channel list items for the redesign, ui-components module, and sample. The list items are not visually complete yet, but they provide enough functionality for hooking click events to complete the navigation flow to the chat/channel/message list screen.

Also contained in this implementation is a slight refactor of the channel list item view holder api. The previous implementation was unusable, as it required clients to provide an instance of a BaseChannelListItemViewHolder. When extending that class, you are required to override setStyle which takes a ChannelListViewStyle as a parameter. However, ChannelListViewStyle is internal to the SDK, and you can't proceed from there.

With the refactor, to specify your own views / viewholders, all you need to do is

1) Set the view holder layout as a ChannelsView attribute.
2) Create a ViewHolder that extends BaseChannelListItemViewHolder.
3) Create a ViewHolder factory that extends BaseChannelViewHolderFactory, and provide the ViewHolder from Step 2.

BREAKING CHANGE: Would require clients to extend BaseChannelViewHolderFactory instead of ChannelViewHolderFactory