Remove unnecessary safe fields in MoshiChatApi
This was added earlier to prevent crashes to a bug that has already been fixed
in the SDK. This reverts to the original lateinit vars that were in this class,
as these values must be initialized before set. Otherwise we're sending blank
values to the backend, which will make the calls fail.