Fix incorrect extension parsing for attachments
Previous method would fail on attachments names that have spaces in them
as it was meant to parse URLs where that's not allowed. All it was really
doing internally was some URL manipulation (removing query params and
fragments, which make no sense for us), plus the logic that it was
now replaced with.