Refactor: Remove unnecessary dark mode parameter from channel and mention list previews
The dark mode parameter in the composable previews for channel and mention lists is not needed anymore, as the `ChatTheme` composable is used directly inside them to handle dark mode. This change removes the extra parameter and makes the previews simpler.
Also removed the `darkMode: Boolean` parameters from snapshot tests.