Provide LocalOnBackPressedDispatcherOwner for snapshot tests
This change provides a fake `OnBackPressedDispatcherOwner` for snapshot tests. This is necessary for composable components that utilize `BackHandler`. The added `FakeBackDispatcherOwner` object implements the required interface and provides a basic lifecycle setup, allowing components like `BackHandler` to function correctly in snapshot testing environments. Also, `LocalOnBackPressedDispatcherOwner` was added to all tests.