Deprecate MessageListHeaderCenterContent and make onClick optional
The `MessageListHeaderCenterContent` composable function in `ChatComponentFactory` has been deprecated. A new version is introduced that makes the `onClick` parameter optional.

This change also updates the `MessageListHeader` and `DefaultMessageListHeaderCenterContent` composables to use the new version of `MessageListHeaderCenterContent`, allowing for a null `onHeaderTitleClick`. The click listener for the header title is now only applied if `onHeaderTitleClick` is not null.