Refactor: Decouple Channel Info top bars from the content
The top bars in `DirectChannelInfoScreen` and `GroupChannelInfoScreen` are now separate from the content and are customizable through the `ChatComponentFactory`.

The `DirectChannelInfoTopBar` and `GroupChannelInfoTopBar` in `ChatComponentFactory` now accept a `ChannelHeaderViewState` as a parameter.

The `GroupChannelInfoTopBar` now displays a "Group info" title along with the channel name. This string has been added to `strings.xml`.