Message members from the member bottom sheet of channel info screen (#5797)
* Fix loading transition

* Clarify that adaptive layout is experimental

* Navigate to distinct channel

* Navigate to new channel

* Refactor: Convert `ChannelInfoMemberViewController.state` to use `stateIn`

This commit refactors the `state` property in `ChannelInfoMemberViewController` to utilize the `stateIn` operator. This simplifies the state management by directly transforming the upstream flow into a `StateFlow`.

The `onChannelInfoData` function is removed as its logic is now incorporated into the `map` operation within the `stateIn` flow.

* Refactor: Use full `cid` instead of `channelId` for channel navigation

The `channelId` parameter, representing only the ID part of a channel identifier, has been replaced with `cid` or `distinctCid`, which represent the full channel identifier (e.g., `messaging:channelId`).

This change affects:
- `ChannelInfoViewEvent.NavigateToChannel`
- `ChannelInfoMemberViewEvent.MessageMember`
- Related event handling in `ChannelInfoViewController` and `ChannelInfoMemberViewController`
- Sample app usage in `ChatsActivity` and `GroupChannelInfoActivity`

This ensures clarity and consistency when referring to channel identifiers.

* Test MessageMember event

* Update queryChannels call in `ChannelInfoMemberViewControllerTest`

The `queryChannels` call in the `ChannelInfoMemberViewControllerTest` has been updated to use a `QueryChannelsRequest` object instead of an `any()` matcher. This ensures that the test accurately reflects the actual call being made by the `ChannelInfoMemberViewController`.

* Refactor: Convert `ChannelInfoMemberViewController.state` to use `stateIn`

This commit refactors the `state` property in `ChannelInfoMemberViewController` to utilize the `stateIn` operator. This simplifies the state management by directly transforming the upstream flow into a `StateFlow`.

The `onChannelInfoData` function is removed as its logic is now incorporated into the `map` operation within the `stateIn` flow.

* Refactor: Navigate to draft channel instead of creating a new direct channel

When messaging a member from the channel info screen, the app will now navigate to a draft channel instead of immediately creating a new direct channel. This allows the user to compose a message before the channel is actually created.

This change also introduces the following new components:
- `DraftChannelViewController`: Manages the state and events for the draft channel screen.
- `DraftChannelViewState`: Represents the state of the draft channel screen.
- `DraftChannelViewAction`: Represents user actions on the draft channel screen.
- `DraftChannelViewEvent`: Represents side effects from the draft channel screen.
- `DraftChannelScreen`: The composable UI for the draft channel.
- `DraftChannelActivity`: The activity that hosts the draft channel screen.

* apiDump

* CHANGELOG