Refactor the XML sample view models and UI components to use ChannelInfoViewController (#5801)
* Introduce ChannelInfo ViewModel, ViewModelFactory and ChannelInfoMember ViewModel, ViewModelFactory to the UI components

* Refactor XML channel info UI for the new view controllers

* Add `SingleLiveEvent` to prevent event re-emission on config changes

This change introduces `SingleLiveEvent`, a `LiveData` variant that only emits new updates after subscription, preventing re-emission of events (like navigation) upon configuration changes.

It also updates `ChannelInfoViewModel` and `ChannelInfoMemberViewModel` to use `asSingleLiveEvent()` for their event streams.

* Add Ban/Unban members in XML sample app

The `BanMember` and `UnbanMember` options were implemented in the `GroupChatInfoMemberOptionsDialogFragment`.

The ban confirmation dialog was added to the `GroupChatInfoFragment`.

* Fix: Stateful option listener called during binding

The `setOnCheckedChangeListener` was being triggered when `binding.optionSwitch.isChecked = item.isChecked` was called during the `bind` process.

This change ensures the listener is only triggered by user interaction by:
1. Temporarily setting the listener to `null` before updating `isChecked`.
2. Restoring the original listener after `isChecked` is updated.

* Refactor: Use shared code in ChatInfoFragment and GroupChatInfoFragment

This commit introduces a new file, `ChatInfoExtensions.kt`, which contains shared code for `ChatInfoFragment` and `GroupChatInfoFragment`.

The following changes were made:

- **Moved `showError` and `showModal` functions to `ChatInfoExtensions.kt`**: These functions are now shared between the two fragments, reducing code duplication.
- **Moved `toChannelInfoItems` function to `ChatInfoExtensions.kt`**: This function is now shared between the two fragments, reducing code duplication.
- **Updated `ChatInfoFragment` and `GroupChatInfoFragment` to use the shared code**: The fragments now use the shared functions from `ChatInfoExtensions.kt`.
- **Made `ChatInfoItem.Option.HideChannel` stateful**: This allows the UI to reflect the current state of the hide channel option.
- **Added `onDismissListener` to `ConfirmationDialogFragment`**: This allows the UI to be updated when a confirmation dialog is dismissed.

* Remove old ViewModels

* apiDump

* Use accent red for banned members

* Remove redundant "Member" role in group info

The "Member" role for users in the chat group info screen was redundant, as it's implied if no other role (like "Owner" or "Moderator") is displayed. This change removes the explicit "Member" label to simplify the UI.

* CHANGELOG

* Introduces experimental classes: `ChannelInfoViewModel`, `ChannelInfoViewModelFactory`, `ChannelInfoMemberViewModel`, and `ChannelInfoMemberViewModelFactory`.

* Remove unused UserData argument from navigation graph

This commit removes the `userData` argument from the `chatInfoFragment` and `chatPreviewFragment` in the main navigation graph as it's no longer used.

* Remove context from `ChatViewModelFactory`

The `Context` parameter was removed from `ChatViewModelFactory` as it was no longer needed for any of its created ViewModels.

* Refactor: Make `HideChannel` option stateless

The `HideChannel` option in `ChatInfoItem` has been changed from a `Stateful` option to a regular `Option`. This simplifies the logic in `ChatInfoFragment` and `GroupChatInfoFragment` by removing the need to manually update the option's checked state when a modal is dismissed.

The click handling for the `HideChannel` option has been updated to reflect this change. It now checks the current `isChecked` state of the option to determine whether to trigger `HideChannelClick` or `UnhideChannelClick`.