Component factories for channel info screens (#5812)
* Factory for `DirectChannelInfoTopBar`

* Introduce `CompoundComponentFactory` for custom component creation using local composition

This commit introduces a public `CompoundComponentFactory` composable function. This function allows users to create a new `ChatComponentFactory` by composing it with the existing one.

This change is useful for:
- Creating custom components.
- Modifying existing components without directly altering the original factory.

The `ChatsActivity` in the sample app and the `ChatsScreen` in the Compose UI library have been updated to utilize this new `CompoundComponentFactory`.

* Remove unnecessary `@Suppress("LongParameterList")` as it is already suppressed in the interface declaration level

* Refactor: Decouple Channel Info top bars from the content

The top bars in `DirectChannelInfoScreen` and `GroupChannelInfoScreen` are now separate from the content and are customizable through the `ChatComponentFactory`.

The `DirectChannelInfoTopBar` and `GroupChannelInfoTopBar` in `ChatComponentFactory` now accept a `ChannelHeaderViewState` as a parameter.

The `GroupChannelInfoTopBar` now displays a "Group info" title along with the channel name. This string has been added to `strings.xml`.

* Refactor: Rename `handleChannelInfoEvents` to `OnChannelInfoEvents`

* Factory for `ChannelInfoOptionItem`

* Remove Default prefix from component names

* Factory for `GroupChannelInfoExpandMemberItem`

* Factory for `ChannelInfoSeparatorItem`

* Factory for `GroupChannelInfoExpandMemberItem`

* Factory for `GroupChannelInfoMemberItem`

* Expose LazyItemScope for ChannelInfoOptionItem, SeparatorItem and ExpandMemberItem

* Factory for `ChannelInfoScreenModal`

* Factory for `ChannelInfoMemberOptionItem`

* Factory for `ChannelInfoMemberInfoModalSheetTopBar`

* Factory for `DirectChannelInfoAvatarContainer`

* CHANGELOG

* apiDump

* merge

* Use UserAvatar from component factory

* Rename GroupChannelInfoExpandMemberItem to GroupChannelInfoExpandMembersItem