Unread message count should not be reset on entering a thread (#5836)
* Filter MessageReadEvent from threads in ChannelLogic

The `MessageReadEvent` was being processed for both parent messages and thread replies. This change ensures that the `MessageReadEvent` is only processed if it's not from a thread (`event.thread == null`). Thread message read events are handled by `QueryThreadsLogic`.

* Add unreadThreads to ChannelUserRead

This commit adds `unreadThreads` to the `ChannelUserRead` model and related entities and mappers.

It also updates `ChannelMutableState.unreadCount` to use `unreadThreads` if available, otherwise it defaults to `unreadMessages`.

* apiDump

* Revert most of the changes and keep only the checking for thread when processing NotificationMarkReadEvent and MessageReadEvent events

* Revert

* CHANGELOG

* Add handling for NotificationMarkUnread in QueryThreadsLogic.kt.

---------

Co-authored-by: PetarVelikov <petar.velikov@getstream.io>