Add long click listener to MapBox and SharedLocationItem
This change introduces a long click listener to the `MapBox` composable and its usage within `SharedLocationItem`.

The `MapBox` now accepts an optional `onLongClick` lambda, which is triggered when the map is long-pressed. This functionality is implemented using `combinedClickable`.

The `SharedLocationItem` has been updated to pass the `onLongItemClick` callback (which was already available for the message item) to the `onLongClick` parameter of `MapBox`. This allows for long-click actions on shared location messages, such as opening message options.