Move MVI classes to ChannelMediaAttachmentsPreviewViewModel
The MVI (Model-View-Intent) classes `ChannelMediaAttachmentsPreviewViewState`, `ChannelMediaAttachmentsPreviewViewAction`, and `ChannelMediaAttachmentsPreviewViewEvent` have been moved into the `ChannelMediaAttachmentsPreviewViewModel.kt` file.

This change also introduces a `ShareLocalFile` event to the `ChannelMediaAttachmentsPreviewViewEvent` sealed interface. This event is used to trigger the sharing of a local file from the `ChannelMediaAttachmentsPreviewScreen`. The `ChannelMediaAttachmentsPreviewViewModel` now uses the application context instead of the regular context.