Save shared images in the Stream cache directory
Instead of `getExternalFilesDir(Environment.DIRECTORY_PICTURES)` or `context.cacheDir`, this change ensures that images shared via the MediaGalleryPreviewActivity are saved into the dedicated Stream cache directory obtained through `StreamFileUtil.getOrCreateStreamCacheDir(context)`.

Additionally, the `MediaGalleryPreviewActivity` and `ChannelMediaAttachmentsPreviewViewModel` now utilize `StreamImageLoader` to load the image as a Bitmap before writing it to a sharable file, specifically for image attachments. This allows for consistent image handling and caching.