Introduce share action to the ChannelMediaAttachmentsPreviewScreen (#5927)
* Add file sharing utility function to ContextUtils

* Refactor attachment sharing logic in `MediaGalleryPreviewActivity` and `MediaGalleryPreviewScreen`.

The main changes include:
- Consolidating image and video sharing into a single `shareAttachment` function.
- Using `StreamFileUtil.writeFileToShareableFile` to prepare attachments for sharing.
- Updating the UI to reflect sharing progress and provide a cancel option.
- Improving the content description for the share icon.
- Making the large file prompt title and message more user-friendly.
- Using constants for file sizes.

* Support images and files in StreamFileUtil.writeFileToShareableFile

* Allow sharing text along with a file

* Introduce sharing functionality for media attachments with error handling

* Add sharing text for media attachments in MediaGalleryPreviewActivity

* detekt

* apiDump

* Move MVI classes to ChannelMediaAttachmentsPreviewViewModel

The MVI (Model-View-Intent) classes `ChannelMediaAttachmentsPreviewViewState`, `ChannelMediaAttachmentsPreviewViewAction`, and `ChannelMediaAttachmentsPreviewViewEvent` have been moved into the `ChannelMediaAttachmentsPreviewViewModel.kt` file.

This change also introduces a `ShareLocalFile` event to the `ChannelMediaAttachmentsPreviewViewEvent` sealed interface. This event is used to trigger the sharing of a local file from the `ChannelMediaAttachmentsPreviewScreen`. The `ChannelMediaAttachmentsPreviewViewModel` now uses the application context instead of the regular context.

* CHANGELOG

* Refactor file operations and add tests for ChannelMediaAttachmentsPreviewViewModel

Refactor file operations in `StreamFileUtil` to be suspend functions and move them to the IO dispatcher.
Introduce `AttachmentFileController` to encapsulate file operations for `ChannelMediaAttachmentsPreviewViewModel`.
Update `MediaGalleryPreviewActivity` and `AttachmentGalleryActivity` to use the new suspend functions.
Add unit tests for `ChannelMediaAttachmentsPreviewViewModel` to cover sharing logic.

* Add fileSize to MediaGalleryPreviewActivityAttachmentState, so that it can be used to check if the media needs to be downloaded before sharing

* Save shared images in the Stream cache directory

Instead of `getExternalFilesDir(Environment.DIRECTORY_PICTURES)` or `context.cacheDir`, this change ensures that images shared via the MediaGalleryPreviewActivity are saved into the dedicated Stream cache directory obtained through `StreamFileUtil.getOrCreateStreamCacheDir(context)`.

Additionally, the `MediaGalleryPreviewActivity` and `ChannelMediaAttachmentsPreviewViewModel` now utilize `StreamImageLoader` to load the image as a Bitmap before writing it to a sharable file, specifically for image attachments. This allows for consistent image handling and caching.

* Re-introduce the deprecated bottom bar component for channel media attachments preview

* The `AttachmentFileController` class was moved from the `ChannelMediaAttachmentsPreviewViewModel.kt` file to its own file in the `io.getstream.chat.android.compose.util` package. This makes it a common class that can be used by other components.
The `MediaGalleryPreviewActivity` was updated to use the new `AttachmentFileController` class.
The `StreamFileUtil.writeImageToSharableFile` function was updated to return a `Result<Uri>` instead of a `Uri?`. This makes the function more robust and easier to use.

* Remove unnecessary fillMaxWidth modifier from ChannelMediaAttachmentsPreviewBottomBar